+++ packages/terra-paginator/src/ControlledPaginator.jsx
@@ -115,15 +115,20 @@ class Paginator extends React.Component {
 
   defaultPaginator() {
     const theme = this.context;
-    const totalPages = calculatePages(this.props.totalCount, this.props.itemCountPerPage);
-    const { selectedPage, intl } = this.props;
+    const {
+      selectedPage,
+      intl,
+      totalCount,
+      itemCountPerPage,
+    } = this.props;
+    const totalPages = calculatePages(totalCount, itemCountPerPage);
     const previousPageIndex = selectedPage === 1 ? 1 : selectedPage - 1;
     const nextPageIndex = selectedPage === totalPages ? totalPages : selectedPage + 1;
 
     const fullView = (
-      <div className={cx('paginator', !this.props.totalCount && 'pageless', theme.className)}>
+      <div className={cx('paginator', !totalCount && 'pageless', theme.className)}>
         {
-          this.props.totalCount && (
+          totalCount && (
           <PaginatorButton
             ariaDisabled={selectedPage === 1}
             ariaLabel={intl.formatMessage({ id: 'Terra.paginator.first' })}
@@ -147,7 +152,7 @@ class Paginator extends React.Component {
           <span className={cx('icon')} />
           {intl.formatMessage({ id: 'Terra.paginator.previous' })}
         </PaginatorButton>
-        {this.props.totalCount && this.buildPageButtons(totalPages, this.handlePageChange)}
+        {totalCount && this.buildPageButtons(totalPages, this.handlePageChange)}
         <PaginatorButton
           ariaDisabled={selectedPage === totalPages}
           ariaLabel={intl.formatMessage({ id: 'Terra.paginator.next' })}
@@ -160,7 +165,7 @@ class Paginator extends React.Component {
           <span className={cx('icon')} />
         </PaginatorButton>
         {
-          this.props.totalCount && (
+          totalCount && (
           <PaginatorButton
             ariaDisabled={selectedPage === totalPages}
             ariaLabel={intl.formatMessage({ id: 'Terra.paginator.last' })}
@@ -181,15 +186,20 @@ class Paginator extends React.Component {
 
   reducedPaginator() {
     const theme = this.context;
-    const totalPages = calculatePages(this.props.totalCount, this.props.itemCountPerPage);
-    const { selectedPage, intl } = this.props;
+    const {
+      selectedPage,
+      intl,
+      totalCount,
+      itemCountPerPage,
+    } = this.props;
+    const totalPages = calculatePages(totalCount, itemCountPerPage);
     const previousPageIndex = selectedPage === 1 ? 1 : selectedPage - 1;
     const nextPageIndex = selectedPage === totalPages ? totalPages : selectedPage + 1;
 
     const reducedView = (
-      <div className={cx('paginator', !this.props.totalCount && 'pageless', theme.className)} role="navigation" aria-label="pagination">
+      <div className={cx('paginator', !totalCount && 'pageless', theme.className)} role="navigation" aria-label="pagination">
         {
-          this.props.totalCount && (
+          totalCount && (
           <PaginatorButton
             ariaDisabled={selectedPage === 1}
             ariaLabel={intl.formatMessage({ id: 'Terra.paginator.first' })}
@@ -213,7 +223,7 @@ class Paginator extends React.Component {
           <span className={cx('icon')} />
           {intl.formatMessage({ id: 'Terra.paginator.previous' })}
         </PaginatorButton>
-        {this.props.totalCount && intl.formatMessage({ id: 'Terra.paginator.pageIndex' }, { pageNumber: selectedPage })}
+        {totalCount && intl.formatMessage({ id: 'Terra.paginator.pageIndex' }, { pageNumber: selectedPage })}
         <PaginatorButton
           ariaDisabled={selectedPage === totalPages}
           ariaLabel={intl.formatMessage({ id: 'Terra.paginator.next' })}
@@ -226,7 +236,7 @@ class Paginator extends React.Component {
           <span className={cx('icon')} />
         </PaginatorButton>
         {
-          this.props.totalCount && (
+          totalCount && (
           <PaginatorButton
             ariaDisabled={selectedPage === totalPages}
             ariaLabel={intl.formatMessage({ id: 'Terra.paginator.last' })}