+++ packages/terra-paginator/src/ControlledProgressivePaginator.jsx
@@ -68,20 +68,25 @@ class ControlledProgressivePaginator extends React.Component {
 
   defaultProgressivePaginator() {
     const theme = this.context;
-    const totalPages = (this.props.totalCount) ? calculatePages(this.props.totalCount, this.props.itemCountPerPage) : 0;
-    const { selectedPage, intl } = this.props;
+    const {
+      selectedPage,
+      intl,
+      totalCount,
+      itemCountPerPage,
+    } = this.props;
+    const totalPages = (totalCount) ? calculatePages(totalCount, itemCountPerPage) : 0;
     const previousPageIndex = selectedPage === 1 ? 1 : selectedPage - 1;
     const nextPageIndex = selectedPage === totalPages ? totalPages : selectedPage + 1;
 
     return (
       <div className={cx('paginator', 'progressive', theme.className)} role="navigation" aria-label="pagination">
         <div>
-          {(this.props.totalCount) ? intl.formatMessage({ id: 'Terra.paginator.pageCount' }, { pageNumber: selectedPage, pageNumberTotal: totalPages })
+          {(totalCount) ? intl.formatMessage({ id: 'Terra.paginator.pageCount' }, { pageNumber: selectedPage, pageNumberTotal: totalPages })
             : intl.formatMessage({ id: 'Terra.paginator.pageIndex' }, { pageNumber: selectedPage })}
         </div>
         <div>
           {
-            this.props.totalCount && (
+            totalCount && (
               <PaginatorButton
                 ariaDisabled={selectedPage === 1}
                 ariaLabel={intl.formatMessage({ id: 'Terra.paginator.first' })}
@@ -117,7 +122,7 @@ class ControlledProgressivePaginator extends React.Component {
             <span className={cx('icon')} />
           </PaginatorButton>
           {
-            (this.props.totalCount) && (
+            (totalCount) && (
               <PaginatorButton
                 ariaDisabled={selectedPage === totalPages}
                 ariaLabel={intl.formatMessage({ id: 'Terra.paginator.last' })}
@@ -137,8 +142,13 @@ class ControlledProgressivePaginator extends React.Component {
 
   reducedProgressivePaginator() {
     const theme = this.context;
-    const totalPages = (this.props.totalCount) ? calculatePages(this.props.totalCount, this.props.itemCountPerPage) : 0;
-    const { selectedPage, intl } = this.props;
+    const {
+      selectedPage,
+      intl,
+      totalCount,
+      itemCountPerPage,
+    } = this.props;
+    const totalPages = (totalCount) ? calculatePages(totalCount, itemCountPerPage) : 0;
     const previousPageIndex = selectedPage === 1 ? 1 : selectedPage - 1;
     const nextPageIndex = selectedPage === totalPages ? totalPages : selectedPage + 1;
 
@@ -146,7 +156,7 @@ class ControlledProgressivePaginator extends React.Component {
       <div className={cx('paginator', theme.className)} role="navigation" aria-label="pagination">
         <div>
           {
-            (this.props.totalCount) && (
+            (totalCount) && (
               <PaginatorButton
                 ariaDisabled={selectedPage === 1}
                 ariaLabel={intl.formatMessage({ id: 'Terra.paginator.first' })}
@@ -172,7 +182,7 @@ class ControlledProgressivePaginator extends React.Component {
           </PaginatorButton>
         </div>
         <div>
-          {(this.props.totalCount) ? intl.formatMessage({ id: 'Terra.paginator.pageCount' }, { pageNumber: selectedPage, pageNumberTotal: totalPages })
+          {(totalCount) ? intl.formatMessage({ id: 'Terra.paginator.pageCount' }, { pageNumber: selectedPage, pageNumberTotal: totalPages })
             : intl.formatMessage({ id: 'Terra.paginator.pageIndex' }, { pageNumber: selectedPage })}
         </div>
         <div>
@@ -188,7 +198,7 @@ class ControlledProgressivePaginator extends React.Component {
             <span className={cx('icon')} />
           </PaginatorButton>
           {
-            (this.props.totalCount) && (
+            (totalCount) && (
               <PaginatorButton
                 ariaDisabled={selectedPage === totalPages}
                 ariaLabel={intl.formatMessage({ id: 'Terra.paginator.last' })}