+++ packages/terra-paginator/src/terra-dev-site/test/paginator/ControlledProgressivePaginatorWithoutTotalCount.test.jsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState } from 'react';
 import Button from 'terra-button';
 import classNames from 'classnames/bind';
 import ControlledProgressivePaginator from '../../../ControlledProgressivePaginator';
@@ -6,31 +6,21 @@ import styles from './ControlledPaginatorTestCommon.module.scss';
 
 const cx = classNames.bind(styles);
 
-class ControlledProgressivePaginatorWithoutTotalCount extends React.Component {
-  constructor(props) {
-    super(props);
+const ControlledProgressivePaginatorWithoutTotalCount = () => {
+  const [currentPage, setCurrentPage] = useState(1);
 
-    this.state = {
-      currentPage: 1,
-    };
+  const changePages = (index) => {
+    setCurrentPage(index);
+  };
 
-    this.changePages = this.changePages.bind(this);
-  }
-
-  changePages(index) {
-    this.setState({ currentPage: index });
-  }
-
-  render() {
-    return (
-      <div className={cx('paginator-wrapper')}>
-        <Button id="button-9" text="Set Page to 9" onClick={() => { this.setState({ currentPage: 9 }); }} />
-        <Button id="button-15" text="Set Page to 15" onClick={() => { this.setState({ currentPage: 15 }); }} />
-        <Button id="button-45" text="Set Page to 45" onClick={() => { this.setState({ currentPage: 45 }); }} />
-        <ControlledProgressivePaginator onPageChange={this.changePages} selectedPage={this.state.currentPage} />
-      </div>
-    );
-  }
-}
+  return (
+    <div className={cx('paginator-wrapper')}>
+      <Button id="button-9" text="Set Page to 9" onClick={() => { setCurrentPage(9); }} />
+      <Button id="button-15" text="Set Page to 15" onClick={() => { setCurrentPage(15); }} />
+      <Button id="button-45" text="Set Page to 45" onClick={() => { setCurrentPage(45); }} />
+      <ControlledProgressivePaginator onPageChange={changePages} selectedPage={currentPage} />
+    </div>
+  );
+};
 
 export default ControlledProgressivePaginatorWithoutTotalCount;